feat(core,script)!: add repeated track sequences
Track sequences can be repeated by prepending each with the
'Core.Track.repeatedSequenceWhere' smart constructor applied to a number
of times that each is to be repeated, and appending each with the
'Core.Track.sequenceEnd' one.

BREAKING CHANGE: The 'Core.Track.eitherSequenceEnd' smart constructor is
not valid, use the 'Core.Track.sequenceEnd' one instead.