feat(core,script)!: add predefined track parts side offsetting
Predefined track parts can be offsetted from the left, the right, or
both by creating them with the 'leftPredefinedPart',
'rightPredefinedPart', or 'middlePredefinedPart' smart constructor,
respectively, applied to a 'Core.Track.Cell' value to offset with and
not empty 'Core.Track.Cell' track lines whose lengths are the same and
not greater than a track width, otherwise they are ignored.

BREAKING CHANGE: The 'Core.Track.predefinedPart' is not a valid smart
constructor, use the 'Core.Track.middlePredefinedPart' one instead.