feat(core,driver): add enemies killing
Enemies can be killed by pressing the 'Swing Left' or 'Swing Right' key
binding (the j and the k key by default) to the right or left of them,
respectively to acquire score bouns ('10' by default). The enemy killing
score bonus can be configured by setting it as a value of the
'enemyKillingScoreBonus' key in a configuration file or the
'--enemy-killing-score-bonus' ('-k') option passed to the command.